+++ src/cli/analyzeCommand.ts
@@ -1,12 +1,11 @@
 import { ConfigManager } from "@config/configManager";
-import { BasicAnalyzer } from "@core/analyzers/basicAnalyzer";
-import { DeepAnalyzer } from "@core/analyzers/deepAnalyzer";
 import { ReportFactory } from "@core/factories/reportFactory";
 import { ComplexityLogger } from "@core/loggers/complexityLogger";
 import { ArgManager } from "@core/managers/argManager";
 import { ReportManager } from "@core/managers/reportManager";
 import { select } from "@inquirer/prompts";
 import { LabInsightArgType } from "@lab-types/arg.type";
+import { ReportFormat } from "@lab-types/report.type";
 import { LoggerUtil } from "@utils/loggerUtil";
 import chalk from "chalk";
 import path from "path";
@@ -56,6 +55,7 @@ export class AnalyzeCommand {
     // Get the stored args as a map
     const args = argManager.getArgs();
     let analysisType: LabInsightArgType = args.type as LabInsightArgType;
+    let outputFormat: string = args.format as ReportFormat;
 
     if (!analysisType) {
       analysisType = await select({
@@ -84,27 +84,61 @@ export class AnalyzeCommand {
         ],
       });
     }
+    if (!outputFormat) {
+      outputFormat = await select({
+        message: "Select the output format for the analysis",
+        choices: [
+          {
+            name: "JSON",
+            value: "json",
+            description: "Generates a JSON report",
+          },
+          {
+            name: "HTML",
+            value: "html",
+            description: "Generates an HTML report",
+          },
+          {
+            name: "CSV",
+            value: "csv",
+            description: "Generates a CSV report",
+            disabled: true,
+          },
+          {
+            name: "XML",
+            value: "xml",
+            description: "Generates an XML report",
+            disabled: true,
+          },
+        ],
+      });
+    }
 
-    await this.startAnalysis(analysisType);
+    const format = outputFormat as ReportFormat;
+
+    await this.startAnalysis(analysisType, format);
   }
 
   /**
    * Initiates the selected type of analysis.
    * @param type LabInsightArgType
    */
-  private async startAnalysis(type: LabInsightArgType): Promise<void> {
+  private async startAnalysis(
+    type: LabInsightArgType,
+    outputFormat: ReportFormat
+  ): Promise<void> {
     switch (type) {
       case "full":
-        await this.performFullAnalysis();
+        await this.performFullAnalysis(outputFormat);
         break;
       case "dependencies":
-        await this.startDependencyAnalysis();
+        await this.startDependencyAnalysis(outputFormat);
         break;
       case "basic":
-        await this.startBasicAnalysis();
+        await this.startBasicAnalysis(outputFormat);
         break;
       case "deep":
-        await this.startDeepAnalysis();
+        await this.startDeepAnalysis(outputFormat);
         break;
       default:
         console.error("Invalid analysis type, please select a valid type");
@@ -112,7 +146,7 @@ export class AnalyzeCommand {
     }
   }
 
-  private async startDependencyAnalysis(): Promise<void> {
+  private async startDependencyAnalysis(format: ReportFormat): Promise<void> {
     try {
       const reportFactory = new ReportFactory();
       await reportFactory.buildReport("dependencies");
@@ -120,7 +154,17 @@ export class AnalyzeCommand {
       const dependencyReport = reportFactory.getDependencyReport();
 
       const reportManager = new ReportManager();
-      reportManager.generateReport("dependencies", "json", dependencyReport);
+      reportManager.generateReport("dependencies", format, dependencyReport);
+
+      const logger = new LoggerUtil();
+
+      logger.spacing();
+
+      console.log(chalk.blue("Enjoying LabInsight ? We value your feedback !"));
+      console.log(
+        chalk.blue("Please leave a review here: ") +
+          chalk.underline("https://forms.gle/uEKij6RwHaucdpQ8A")
+      );
     } catch (error) {
       console.error(
         "An error occurred while performing dependency analysis",
@@ -132,7 +176,7 @@ export class AnalyzeCommand {
   /**
    * Starts the basic analysis.
    */
-  private async startBasicAnalysis(): Promise<void> {
+  private async startBasicAnalysis(format: ReportFormat): Promise<void> {
     try {
       const reportFactory = new ReportFactory();
       await reportFactory.buildReport("basic");
@@ -140,7 +184,17 @@ export class AnalyzeCommand {
       const basicReport = reportFactory.getBasicReport();
 
       const reportManager = new ReportManager();
-      reportManager.generateReport("basic", "json", basicReport);
+      reportManager.generateReport("basic", format, basicReport);
+
+      const logger = new LoggerUtil();
+
+      logger.spacing();
+
+      console.log(chalk.blue("Enjoying LabInsight ? We value your feedback !"));
+      console.log(
+        chalk.blue("Please leave a review here: ") +
+          chalk.underline("https://forms.gle/uEKij6RwHaucdpQ8A")
+      );
     } catch (error) {
       console.error("An error occurred while performing basic analysis", error);
     }
@@ -149,7 +203,7 @@ export class AnalyzeCommand {
   /**
    * Starts the deep analysis.
    */
-  private async startDeepAnalysis(): Promise<void> {
+  private async startDeepAnalysis(format: ReportFormat): Promise<void> {
     try {
       const reportFactory = new ReportFactory();
       await reportFactory.buildReport("deep");
@@ -160,7 +214,17 @@ export class AnalyzeCommand {
       complexityLogger.logCodeComplexityResult(deepReport.complexity);
 
       const reportManager = new ReportManager();
-      reportManager.generateReport("deep", "json", deepReport);
+      reportManager.generateReport("deep", format, deepReport);
+
+      const logger = new LoggerUtil();
+
+      logger.spacing();
+
+      console.log(chalk.blue("Enjoying LabInsight ? We value your feedback !"));
+      console.log(
+        chalk.blue("Please leave a review here: ") +
+          chalk.underline("https://forms.gle/uEKij6RwHaucdpQ8A")
+      );
     } catch (error) {
       console.error("An error occurred while performing deep analysis", error);
     }
@@ -169,14 +233,14 @@ export class AnalyzeCommand {
   /**
    * Performs a full analysis (dependencies + basic + deep).
    */
-  private async performFullAnalysis(): Promise<void> {
+  private async performFullAnalysis(format: ReportFormat): Promise<void> {
     const logger = new LoggerUtil();
     const reportsFolder = path.join(process.cwd(), "reports");
 
     try {
-      await this.startDependencyAnalysis();
-      await this.startBasicAnalysis();
-      await this.startDeepAnalysis();
+      await this.startDependencyAnalysis(format);
+      await this.startBasicAnalysis(format);
+      await this.startDeepAnalysis(format);
 
       logger.spacing();
 
@@ -185,6 +249,14 @@ export class AnalyzeCommand {
         chalk.greenBright(
           `Reports are available in ${chalk.bold(reportsFolder)}`
         )
+      );
+
+      logger.spacing();
+
+      console.log(chalk.blue("Enjoying LabInsight ? We value your feedback !"));
+      console.log(
+        chalk.blue("Please leave a review here: ") +
+          chalk.underline("https://forms.gle/uEKij6RwHaucdpQ8A")
       );
     } catch (error) {
       console.error("An error occurred while performing full analysis", error);