+++ src/languages/python/analyzers/pyDeepAnalyzer.ts
@@ -1,0 +1,5 @@
+export class PyDeepAnalyzer {
+  constructor() {
+    // TODO: Implement this class
+  }
+}